+++ cc/920_number_of_music_playlists.cc
@@ -26,8 +26,8 @@ class Solution {
         if (n == 0 || goal == 0) return 0;
         if (dp.at(n).at(goal) != -1) return dp.at(n).at(goal);
         ll const pick{Solve(n - 1, goal - 1, k, dp) * n};
-        ll const notpick{Solve(n, goal - 1, k, dp) * max(n - k, 0)};
-        return dp.at(n).at(goal) = (pick + notpick) % m_mod;
+        ll const notpick{Solve(n, goal - 1, k, dp) * static_cast<ll>(max(n - k, 0))};
+        return dp.at(n).at(goal) = static_cast<int>((pick + notpick) % m_mod);
     }
 
     int NumMusicPlaylists(int n, int goal, int k) {
